# tests/api-docs/openapi_v4.yaml
openapi: 3.0.0
info:
  title: "Trabalho PSD-08"
  version: v4
servers: 
  - url: http://localhost:5010
  - url: http://localhost:5020
...
paths:
...
  /eecs/saida:
    get:
      tags:
      - "EECS"
      description: Apresentar todas as saídas
      responses:
        200:
          description: todas as saídas
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/objSaidaResponse'
    post:
      tags:
      - "EECS"
      description: Registar uma nova saída
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/objSaidaRequest'
      responses:
        '201':
          description: Nova saída efetuada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/objSaidaResponse'
...
components:
  schemas:
...
    objSaidaRequest:
      type: object
      properties:
        matriculaSaidas: 
          type: string 
    objSaidaResponse:
      type: object
      properties:
        _id:
          type: string
        matriculaSaidas:
          type: string
        data: 
          type: string
          format: date
        Created_data: 
          type: string
          format: date-time
...